test class updated and AnnualServiceClass updated

+++ src/main/java/com/acoustic/repository/AnnualNetRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.acoustic.entity.AnnualNet;
 
 @Repository
-public interface AnnualNetRepository extends JpaRepository<AnnualNet, Long> {
+public interface AnnualNetRepository extends JpaRepository<AnnualNet, Integer> {
